NGL not sure how this worked before now, must have had an extra render target I wasn't managing properly. - Added a new framebuffer for game only, rather than having a poorly tracked render target? Not sure
Just fixing issues with non-editor rendering

+++ Flow/Source/Flow/Rendering/Core/Bindables/Bindable.h
@@ -28,7 +28,7 @@ namespace Bindables
 
 		// Private Variables ///////////////////////////////////////////////////////////
 
-		HashString	m_id = 0;
+		HashString m_id = (size_t)0;
 	};
 
 	inline HashString Bindable::GetID()
